FIX: cannot recover temperature for hotend
    Because the parameters for thermalManager.setTargetHotend() is
    reversed. Also optimize the recovery process.

+++ Marlin/src/module/PowerPanic.h
@@ -24,7 +24,7 @@ typedef struct __attribute__((aligned (4)))
 	// checksum of this section
 	uint32_t CheckSum;
 	// temperature of extrucders
-	float HeaterTamp[PP_HEATER];
+	int16_t HeaterTamp[PP_HEATER];
 	// speed of work
 	float PrintFeedRate;
 	// speed of travel
@@ -35,7 +35,7 @@ typedef struct __attribute__((aligned (4)))
 	float laser_percent;
 	uint16_t laser_pwm;
 	// target temperature of heat bed
-	float BedTamp;
+	int16_t BedTamp;
 	// position of stepper on last move
 	float PositionData[NUM_AXIS];
 	// position shift between home offset and workspace offset
